# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ThePower LMS API
  description: API multi-tenant para plataforma LMS B2B con FUNDAE compliance
  version: 1.0.0
  contact:
    name: ThePower Education
servers:
  - url: https://api.thepower.education/v1
    description: Production
  - url: https://api.staging.thepower.education/v1
    description: Staging
security:
  - BearerAuth: []
  - OAuth2: [read, write]
...
paths:
...
  /tenants/{tenantId}/settings:
    get:
      tags: [Tenants]
      summary: Obtener configuración del tenant
      operationId: getTenantSettings
      parameters:
        - $ref: '#/components/parameters/tenantId'
      responses:
        '200':
          description: Configuración del tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantSettings'
    put:
      tags: [Tenants]
      summary: Actualizar configuración del tenant
      operationId: updateTenantSettings
      parameters:
        - $ref: '#/components/parameters/tenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantSettingsUpdate'
      responses:
        '200':
          description: Configuración actualizada
...
  /users:
    get:
      tags: [Users]
      summary: Listar usuarios del tenant
      operationId: listUsers
      parameters:
        - $ref: '#/components/parameters/pageSize'
        - $ref: '#/components/parameters/cursor'
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, pending]
      responses:
        '200':
          description: Lista paginada de usuarios
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserList'
  /users/import:
    post:
      tags: [Users]
      summary: Importar usuarios en bulk (CSV/Excel)
      operationId: importUsers
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                validateOnly:
                  type: boolean
                  default: false
      responses:
        '202':
          description: Importación iniciada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportJob'
  /users/{userId}/roles:
    put:
      tags: [Users]
      summary: Asignar roles a usuario
      operationId: setUserRoles
      parameters:
        - $ref: '#/components/parameters/userId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                roleIds:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        '200':
          description: Roles asignados
  /users/deactivate:
    post:
      tags: [Users]
      summary: Desactivar usuarios en bulk
      operationId: deactivateUsers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userIds:
                  type: array
                  items:
                    type: string
                    format: uuid
                reason:
                  type: string
      responses:
        '200':
          description: Usuarios desactivados

  # === COURSES ===
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    OAuth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: /oauth/token
          scopes:
            read: Read access
            write: Write access
  parameters:
    tenantId:
      name: tenantId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    userId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    pageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
    cursor:
      name: cursor
      in: query
      schema:
        type: string
  schemas:
    TenantSettings:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        domain:
          type: string
        cif:
          type: string
        branding:
          type: object
          properties:
            logoUrl:
              type: string
            primaryColor:
              type: string
            secondaryColor:
              type: string
        featureFlags:
          type: object
          additionalProperties:
            type: boolean
        gamificationConfig:
          $ref: '#/components/schemas/GamificationConfig'
        notificationConfig:
          type: object
    TenantSettingsUpdate:
      type: object
      properties:
        name:
          type: string
        branding:
          type: object
        featureFlags:
          type: object
        gamificationConfig:
          $ref: '#/components/schemas/GamificationConfig'
        notificationConfig:
          type: object
    GamificationConfig:
      type: object
      properties:
        leaderboardEnabled:
          type: boolean
        leaderboardScopes:
          type: array
          items:
            type: string
            enum: [global, department, team]
        optOutEnabled:
          type: boolean
        xpRules:
          type: object
        antiCheatEnabled:
          type: boolean
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        dni:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        status:
          type: string
          enum: [active, inactive, pending]
        roles:
          type: array
          items:
            type: string
    UserList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
        nextCursor:
          type: string
        totalCount:
          type: integer
    ImportJob:
      type: object
      properties:
        jobId:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, processing, completed, failed]
        totalRows:
          type: integer
        processedRows:
          type: integer
        errors:
          type: array
          items:
            type: object
...