# services/iam-service/src/routes/users.ts
/**
 * User Routes
 * CRUD, bulk import, role assignment
 */

import { Hono } from 'hono'
import { zValidator } from '@hono/zod-validator'
import { z } from 'zod'
import { prisma } from '../lib/prisma'
import bcrypt from 'bcryptjs'
import { createClient } from '@supabase/supabase-js'

const userRoutes = new Hono()

// Schemas
const createUserSchema = z.object({
    email: z.string().email(),
    dni: z.string().optional(),
    firstName: z.string().min(1).max(100),
    lastName: z.string().min(1).max(100),
    password: z.string().min(8).optional(),
    department: z.string().optional(),
    managerId: z.string().uuid().optional(),
    roleIds: z.array(z.string().uuid()).optional(),
    role: z.string().optional(),
    status: z.string().optional(),
    client_id: z.string().optional()
})

const updateUserSchema = createUserSchema.partial()

const bulkDeactivateSchema = z.object({
    userIds: z.array(z.string().uuid()).min(1),
    reason: z.string().optional()
})

// POST /api/users - Create User
userRoutes.post('/', zValidator('json', createUserSchema), async (c) => {
    const body = c.req.valid('json')
    const tenantId = c.get('tenantId') as string || 'a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11' // Fallback for prototype

    try {
        const user = await prisma.user.create({
            data: {
                email: body.email,
                firstName: body.firstName,
                lastName: body.lastName,
                jobTitle: body.department, // Map department to job_title
                role: body.role || 'student',
                status: body.status || 'active',
                tenantId: body.client_id || tenantId,
                passwordHash: body.password ? bcrypt.hashSync(body.password, 10) : undefined
            } as any // Bypass strict type check if needed for merged schema
        })

        return c.json(user, 201)
    } catch (error: any) {
        console.error("Prisma Create User Error:", error)
        return c.json({
            error: error.message,
            stack: error.stack,
            meta: error.meta,
            code: error.code
        }, 500)
    }
})

// GET /api/users/debug-test
userRoutes.get('/debug-test', async (c) => {
    const sbUrl = process.env.SUPABASE_URL || ''
    const sbKey = process.env.SUPABASE_ANON_KEY || ''
    const supabase = createClient(sbUrl, sbKey)

    const { data, error } = await supabase.from('users').select('*').limit(5)

    // Explicitly check count too
    const { count } = await supabase.from('users').select('*', { count: 'exact', head: true })

    return c.json({
        message: "Debug Probe",
        sbUrl,
        sbKeyLen: sbKey?.length,
        userCount: data?.length,
        totalCount: count,
        firstUser: data?.[0],
        error
    })
})

// GET /api/users
userRoutes.get('/', async (c) => {
    const tenantId = c.get('tenantId') as string
    const { status, department, cursor, pageSize } = c.req.query()

    try {
        const where: any = {}
        if (tenantId && tenantId.length > 10) where.tenantId = tenantId
        if (status) where.status = status

        // Use jobTitle for department filtering
        if (department) where.jobTitle = { contains: department, mode: 'insensitive' }

        const users = await prisma.user.findMany({
            where,
            orderBy: { createdAt: 'desc' },
            take: 100 // Limit for prototype
        })

        return c.json({
            data: users.map((u: any) => ({
                id: u.id,
                email: u.email || `${u.firstName?.toLowerCase()}.${u.lastName?.toLowerCase()}@demo.com`,
                first_name: u.firstName, // Snake case for frontend
                last_name: u.lastName,
                firstName: u.firstName, // Camel case
                lastName: u.lastName,
                status: u.status || 'active',
                department: u.jobTitle || 'General',
                role: u.role || 'student',
                roles: u.role ? [u.role] : ['student'],
                clients: { name: 'Inditex Group' }, // Mocked
                createdAt: u.createdAt
            }))
        })
    } catch (error: any) {
        console.error("Prisma Get Users Error:", error)
        return c.json({ error: error.message }, 500)
    }
})

// GET /api/users/:id
userRoutes.get('/:id', async (c) => {
    const tenantId = c.get('tenantId') as string
    const { id } = c.req.param()

    const user = await prisma.user.findFirst({
        where: { id, tenantId }, // deletedAt removed
        include: {
            // userRoles: { include: { role: true } } // Relation removed
        }
    })

    if (!user) {
        return c.json({ error: 'User not found' }, 404)
    }

    const u = user as any
    return c.json({
        id: u.id,
        email: `${u.firstName.toLowerCase()}.${u.lastName.toLowerCase()}@demo.com`,
        dni: u.dni || null,
        firstName: u.firstName,
        lastName: u.lastName,
        avatarUrl: u.avatarUrl,
        status: 'active',
        department: u.jobTitle || 'General',
        managerId: u.managerId,
        preferences: u.preferences,
        roles: u.role ? [u.role] : ['student'],
        lastLoginAt: u.lastLoginAt,
        createdAt: u.createdAt
    })
})

// PUT /api/users/:id - Update user details
userRoutes.put('/:id', zValidator('json', updateUserSchema), async (c) => {
    const { id } = c.req.param()
    const body = c.req.valid('json')

    // Supabase Client
    const sbUrl = process.env.SUPABASE_URL || ''
    const sbKey = process.env.SUPABASE_ANON_KEY || ''
    const supabase = createClient(sbUrl, sbKey)

    const updateData: any = {}
    if (body.firstName) updateData.first_name = body.firstName
    if (body.lastName) updateData.last_name = body.lastName
    if (body.department) updateData.job_title = body.department // mapping department to job_title for now
    if (body.status) updateData.status = body.status
    if (body.client_id) updateData.client_id = body.client_id

    const { data, error } = await supabase
        .from('users')
        .update(updateData)
        .eq('id', id)
        .select()
        .single()

    if (error) {
        return c.json({ error: error.message }, 500)
    }

    return c.json(data)
})

// DELETE /api/users/:id - Delete user
userRoutes.delete('/:id', async (c) => {
    const { id } = c.req.param()

    const sbUrl = process.env.SUPABASE_URL || ''
    const sbKey = process.env.SUPABASE_ANON_KEY || ''
    const supabase = createClient(sbUrl, sbKey)

    const { error } = await supabase
        .from('users')
        .delete()
        .eq('id', id)

    if (error) {
        return c.json({ error: error.message }, 500)
    }

    return c.json({ success: true })
})

// POST /api/users/jobs/check-inactivity
userRoutes.post('/jobs/check-inactivity', async (c) => {
    const sbUrl = process.env.SUPABASE_URL || ''
    const sbKey = process.env.SUPABASE_ANON_KEY || ''
    const supabase = createClient(sbUrl, sbKey)

    // Calculate 7 days ago
    const sevenDaysAgo = new Date()
    sevenDaysAgo.setDate(sevenDaysAgo.getDate() - 7)

    // Find users active > 7 days ago OR users with null last_active_at (never active?)
    // Logic: last_active_at < sevenDaysAgo AND status != 'inactive'

    // 1. Get inactive users
    const { data: usersToUpdate, error: fetchError } = await supabase
        .from('users')
        .select('id')
        .lt('last_active_at', sevenDaysAgo.toISOString())
        .neq('status', 'inactive')

    if (fetchError) {
        return c.json({ error: fetchError.message }, 500)
    }

    if (!usersToUpdate || usersToUpdate.length === 0) {
        return c.json({ message: "No inactive users found", count: 0 })
    }

    // 2. Update them
    const ids = usersToUpdate.map(u => u.id)
    const { error: updateError } = await supabase
        .from('users')
        .update({ status: 'inactive' })
        .in('id', ids)

    if (updateError) {
        return c.json({ error: updateError.message }, 500)
    }

    return c.json({
        message: "Inactivity check completed",
        deactivatedCount: ids.length,
        ids
    })
})




// PUT /api/users/:id/roles
userRoutes.put('/:id/roles', async (c) => {
    const tenantId = c.get('tenantId') as string
    const grantedBy = c.get('userId') as string
    const { id } = c.req.param()
    const { roleIds } = await c.req.json()

    const user = await prisma.user.findFirst({
        where: { id, tenantId } // deletedAt removed
    })

    if (!user) {
        return c.json({ error: 'User not found' }, 404)
    }

    // Replace all roles
    await prisma.$transaction([
        prisma.userRole.deleteMany({ where: { userId: id } }),
        prisma.userRole.createMany({
            data: roleIds.map((roleId: string) => ({
                userId: id,
                roleId,
                grantedBy
            }))
        })
    ])

    return c.json({ success: true })
})

// POST /api/users/deactivate
userRoutes.post('/deactivate', zValidator('json', bulkDeactivateSchema), async (c) => {
    const tenantId = c.get('tenantId') as string
    const body = c.req.valid('json')

    const result = await prisma.user.updateMany({
        where: {
            id: { in: body.userIds },
            tenantId,
            // deletedAt: null
        },
        data: {
            // status: 'inactive', // status removed from schema
            updatedAt: new Date()
        }
    })

    return c.json({
        success: true,
        deactivatedCount: result.count
    })
})

// POST /api/users/import
userRoutes.post('/import', async (c) => {
    const tenantId = c.get('tenantId') as string || 'a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11'
    const userId = c.get('userId') as string || 'system'

    // For now, accept JSON array (CSV parsing would be added)
    const body = await c.req.json()
    const { users, validateOnly = false } = body

    const errors: any[] = []
    const validUsers: any[] = []

    // Validate each user
    for (let i = 0; i < users.length; i++) {
        const user = users[i]
        const rowErrors: string[] = []

        if (!user.email || !z.string().email().safeParse(user.email).success) {
            rowErrors.push('Invalid email')
        }
        if (!user.firstName) rowErrors.push('First name required')
        if (!user.lastName) rowErrors.push('Last name required')

        if (rowErrors.length) {
            errors.push({ row: i + 1, errors: rowErrors })
        } else {
            validUsers.push(user)
        }
    }

    if (validateOnly) {
        return c.json({
            totalRows: users.length,
            validRows: validUsers.length,
            errorRows: errors.length,
            errors
        })
    }

    // Create import job
    const job = await prisma.importJob.create({
        data: {
            tenantId,
            fileName: 'bulk_import.json',
            totalRows: users.length,
            createdBy: userId,
            status: 'processing'
        }
    })

    // Process valid users (would be async in production)
    let successCount = 0
    for (const userData of validUsers) {
        try {
            await prisma.user.create({
                data: {
                    tenantId,
                    email: userData.email,
                    firstName: userData.firstName,
                    lastName: userData.lastName,
                    jobTitle: userData.department,
                    role: 'student',
                    status: 'active'
                } as any
            })
            successCount++
        } catch (e) {
            errors.push({ email: userData.email, error: 'Duplicate or invalid' })
        }
    }

    // Update job status
    await prisma.importJob.update({
        where: { id: job.id },
        data: {
            status: 'completed',
            processedRows: users.length,
            successRows: successCount,
            errorRows: errors.length,
            errors,
            completedAt: new Date()
        }
    })

    return c.json({
        jobId: job.id,
        status: 'completed',
        totalRows: users.length,
        successRows: successCount,
        errorRows: errors.length,
        errors
    }, 202)
})

export { userRoutes }